list of news is working, hella warnings in console
