--- src/Post.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Post = ({item}) => {
-  return (
-    <div>News here</div>
-  )
-}
-
-export default Post
