+++ src/Posts.js
@@ -1,19 +1,29 @@
 import React from 'react'
-import Post from './Post'
 
-const Posts = (props) => {
-  const newsItems = props.news.map((item) => {
-    return <Post key={item.id} name={item.name} description={item.description} />
-  });
+const Post = () => {
+  const news = [
+    { id: 1, name: 'Great news!', description: 'This is an example of great news.' },
+    { id: 2, name: 'Okay news', description: 'Nothing out of the ordinary here.' },
+    { id: 3, name: 'Bad news', description: 'Well, this just sucks!' }
+  ]
+
+  console.log(news)
 
   return(
     <div className="posts">
       <h2>Here are all the news!</h2>
       <ul>
-        { newsItems }
+        { news.map((item) => {
+          return (
+            <li>
+              <h3>{item.name}</h3>
+              <p>{item.description}</p>
+            </li>
+          )
+        }) }
       </ul>
     </div>
   )
 }
 
-export default Posts
+export default Post
